+++ phase/reset_workers.go
@@ -10,42 +10,44 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-// UninstallWorkers uninstalls k0s from the workers
-type UninstallWorkers struct {
+// ResetControllers phase removes workers marked for reset from the kubernetes cluster
+// and resets k0s on the host
+type ResetWorkers struct {
 	GenericPhase
 
-	NoDrain bool
+	NoDrain  bool
+	NoDelete bool
 
 	hosts  cluster.Hosts
 	leader *cluster.Host
 }
 
 // Title for the phase
-func (p *UninstallWorkers) Title() string {
-	return "Uninstall workers"
+func (p *ResetWorkers) Title() string {
+	return "Reset workers"
 }
 
 // Prepare the phase
-func (p *UninstallWorkers) Prepare(config *v1beta1.Cluster) error {
+func (p *ResetWorkers) Prepare(config *v1beta1.Cluster) error {
 	p.Config = config
 	p.leader = p.Config.Spec.K0sLeader()
 
 	var workers cluster.Hosts = p.Config.Spec.Hosts.Workers()
 	log.Debugf("%d workers in total", len(workers))
 	p.hosts = workers.Filter(func(h *cluster.Host) bool {
-		return h.Uninstall
+		return h.Reset
 	})
-	log.Debugf("UninstallWorkers phase prepared, %d workers will be uninstalled", len(p.hosts))
+	log.Debugf("ResetWorkers phase prepared, %d workers will be reset", len(p.hosts))
 	return nil
 }
 
-// ShouldRun is true when there are workers that needs to be uninstalled
-func (p *UninstallWorkers) ShouldRun() bool {
+// ShouldRun is true when there are workers that needs to be reset
+func (p *ResetWorkers) ShouldRun() bool {
 	return len(p.hosts) > 0
 }
 
 // CleanUp cleans up the environment override files on hosts
-func (p *UninstallWorkers) CleanUp() {
+func (p *ResetWorkers) CleanUp() {
 	for _, h := range p.hosts {
 		if len(h.Environment) > 0 {
 			if err := h.Configurer.CleanupServiceEnvironment(h, h.K0sServiceName()); err != nil {
@@ -56,7 +58,7 @@ func (p *UninstallWorkers) CleanUp() {
 }
 
 // Run the phase
-func (p *UninstallWorkers) Run() error {
+func (p *ResetWorkers) Run() error {
 	return p.hosts.ParallelEach(func(h *cluster.Host) error {
 		log.Debugf("%s: draining node", h)
 		if !p.NoDrain {
@@ -71,12 +73,14 @@ func (p *UninstallWorkers) Run() error {
 		log.Debugf("%s: draining node completed", h)
 
 		log.Debugf("%s: deleting node...", h)
-		if err := p.leader.DeleteNode(&cluster.Host{
-			Metadata: cluster.HostMetadata{
-				Hostname: h.Metadata.Hostname,
-			},
-		}); err != nil {
-			log.Warnf("%s: failed to delete node: %s", h, err.Error())
+		if !p.NoDelete {
+			if err := p.leader.DeleteNode(&cluster.Host{
+				Metadata: cluster.HostMetadata{
+					Hostname: h.Metadata.Hostname,
+				},
+			}); err != nil {
+				log.Warnf("%s: failed to delete node: %s", h, err.Error())
+			}
 		}
 		log.Debugf("%s: deleting node", h)
 
@@ -110,7 +114,7 @@ func (p *UninstallWorkers) Run() error {
 		}
 		log.Debugf("%s: removing config completed", h)
 
-		log.Infof("%s: uninstalled", h)
+		log.Infof("%s: reset", h)
 		return err
 	})
 }
